Add directory entries for media_dirs
Insert each media_dir as a directory entry to the browse dir
instead of directly adding it's contents to the browse dir.
Then scan the media_dir with it's directory entry as the parent.
That way, the content will be added to the directory entry,
not the browse dir. Hence, the different media_dirs show up
as folders and the contents don't mix up.

modified:   scanner.c